Training now wroks with bf16 and flash_attn can be explicitly forced to

+++ py4cast/observer.py
@@ -185,7 +185,7 @@ class PredictionTimestepPlot(MapPlot):
                 plot_prediction(
                     pred_t[:, :, var_i],
                     target_t[:, :, var_i],
-                    obj.interior_2d[:, :, 0],
+                    obj.interior_2d[:, :, 0].type(torch.float32),
                     title=f"{var_name} ({var_unit}), "
                     f"t={t_i} ({obj.hparams['hparams'].dataset_info.step_duration*t_i} h)",
                     vrange=var_vrange,
@@ -240,7 +240,7 @@ class PredictionEpochPlot(MapPlot):
             plot_prediction(
                 pred_t[:, :, var_i],
                 target_t[:, :, var_i],
-                obj.interior_2d[:, :, 0],
+                obj.interior_2d[:, :, 0].type(torch.float32),
                 title=f"{var_name} ({var_unit}), "
                 f"t={max_step} ({leadtime} h) - epoch {obj.current_epoch}",
                 vrange=var_vrange,
@@ -393,7 +393,7 @@ class SpatialErrorPlot(ErrorObserver):
             loss_map_figs = [
                 plot_spatial_error(
                     loss_map,
-                    obj.interior_2d[:, :, 0],
+                    obj.interior_2d[:, :, 0].type(torch.float32),
                     title=f"{self.prefix} loss, t={t_i} ({obj.hparams['hparams'].dataset_info.step_duration*t_i} h)",
                     domain_info=obj.hparams["hparams"].dataset_info.domain_info,
                 )
